+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 
 ### Changed
 
+- **BREAKING**: Rename schemes `system-version` property to
+  `supports.styling-spec` as per spec `0.2.0-beta2`
 - **BREAKING**: Tinted8 scheme `family`, `style`, and `variant` now live under
   the `meta` object, aligning with spec 0.2.0.
 - Improve errors and messages around missing/invalid template config and